// src/app/videos/models/call-action.ts
export class CallAction {
    isOverlay: boolean;
    email_id: string;
    firstName: string;
    lastName: string;
    showOverLay: boolean;
    isSkipChecked: boolean;
    isFistNameChecked: boolean;
    videoOverlaySubmit: string;
    overLayValue: string;
    startCalltoAction: boolean;
    endCalltoAction: boolean;
 }